move ShortNumberCounter to counters.tsx as GenericCounterRenderer
